Launcher

Launcher

Launcher is the Julia package (sorry, I really, really like writing Julia code) for handling the launching of containers, aggregation of results, binding containers with relevant datasets, and generally making sure everything is working correctly. Documentation for this package can be found in this section.

The functionality provided by this model is very straightforward and can probably be ported to another language if needed.

Note that Launcher is built on top of two other packages:

DockerX - Package for interacting with the Docker API.

MemSnoop - Package for tracking the memory usage patterns of applications on the Linux operating system.

These two packages are still works in progress and documentation on them is forthcoming. However, I plan on registering at least DockerX and probably MemSnoop as well as soon as I take the time to get them production ready.

Temporary Documentation

Base.run — Function.

run([f::Function], work::AbstractWorkload; showlog = false, kw...)

Create and launch a container from work with

container = create(work; kw...)

Start the container and then call f(container). If f is not given, then attach to the container's stdout.

This function ensures that containers are stopped and cleaned up in case something goes wrong.

If showlog = true, send the container's log to stdout when the container stops.

Examples

Using Julia's do syntax to perform a stack based analysis

tracker = run(TestWorkload()) do container
    trackstack(getpid(container))
end

source

Launcher.AbstractWorkload — Type.

Abstract supertype for workloads. Concrete subtypes should be implemented for each workload desired for analysis.

source

Launcher.startfile — Function.

startfile(work::AbstractWorkload, ::Type{OnHost}) -> String

Return the path of the entrypoint file of work on the host machine.

startfile(work::AbstractWorkload, ::Type{OnContainer}) -> String

Return the path of the entrypoint file of work on the Docker Container.

source

Launcher.runcommand — Function.

runcommand(work::AbstractWorkload) -> Cmd

Return the Docker Container entry command for work.

source

Launcher.create — Function.

create(work::AbstractWorkload; kw...) -> Container

Create a Docker Container for work, with optional keyword arguments. Concrete subtypes of AbstractWorkload must define this method and perform all the necessary steps to creating the Container. Note that the container should just be created by a call to DockerX.create_container, and not actually started.

Keyword arguments supported by work should be included in that types documentation.

source