Keras Models

Keras Models

Cifar Cnn

A simple CNN for training on the cifar-10 dataset. This model is small enough that a couple epochs of training takes a reasonably short amount of time, even when snooping memory.

File name: /workloads/keras/cifar_cnn.py

Container entry point: /home/startup/cifar_cnn.py

Dataset: cifar-10-batches-py.tar.gz (https://www.cs.toronto.edu/~kriz/cifar-10-python.tar.gz)

Endpoint for dataset in container: /home/user/.keras/datasets/cifar-10-batches-py.tar.gz. If dataset doesn't exist, it will automatically be downloaded. However, this can take a while and is a bit rude to the site hosting the dataset.

Script Arguments:

--batchsize [size] : Configure the batch size for training.

--epochs [n] : Train for n epochs

--abort : Import the keras and tensorflow libraries and then exit. Used for testing the overhead of code loading.

Launcher Docs:

Launcher.CifarCnn — Type.

Workload object for the Keras Cifar10 cnn. Build type using keyword constructors.

Fields

args :: NamedTuple - Arguments to pass to the startup script (see docs). Default: NamedTuple()

interactive :: Bool - If set to true, the container will launch into /bin/bash instead of Python. Used for debugging the container. Default: false.

create keywords

cpuSets = "" - The CPU sets on which to run the workload. Defaults to all processors. Examples: "0", "0-3", "1,3".

source