Tensorflow Models

Tensorflow Models

Resnet

A model for ResNet that can be trained on either Cifar or ImageNet, though as of now only ImageNet is supported.

File: /workloads/tensorflow/official/resnet/imagenet_main.py

Container entry point: /models/official/resnet/imagenet_main.py

Dataset:

Volume Binds:

Script Arguments

--batch_size=size : Configure batch size. Default: 32

--resnet_size=size : Define the version of ResNet to use. Choices: 18, 34, 50, 101, 152, 200. Default: 50

--train_epochs=N : Number of epochs to train for. Default: 90

--data_dir=path : Path (inside the container) to the data directory. Default provided by Launcher. NOTE: If this is set to something besides /imagenet - things will probably break horribly.

Launcher Docs

Launcher.ResnetTF — Type.

Struct representing parameters for launching the Tensorflow Official Resnet Model on the Imagenet training set. Construct type using a key-word constructor

Fields

args::NamedTuple - Arguments passed to the Keras Python script that creates and trains Resnet.

interactive::Bool - Set to true to create a container that does not automatically run Resnet when launched. Useful for debugging what's going on inside the container.

create keywords

memory::Union{Nothing, Int} - The amount of memory to assign to this container. If this value is nothing, the container will have access to all system memory. Default: nothing.

cpuSets = "" - The CPU sets on which to run the workload. Defaults to all processors. Examples: "0", "0-3", "1,3".

source

Changes Made to imagenet_main.py

Lines 42-47: Made script expect training and validation files to be in train and validation directories respectively whereas the original expected both to be in the same directory.

Aditionally, made the _NUM_TRAIN_FILES and _NUM_VALIDATION_FILES be assigned to the number of files in these directories.

This allows us to operate on a subset of the ImageNet data by just pointing to another folder.

Also hardcoded _DATA_DIR to /imagenet to allow this to take place. This limits the migratability of this project outside of docker, but we'll deal with that when we need to.